comment out stuff that has bad dependencies
for some reason sklearn will not import in my production environment.  I am temporarily commenting things that depend on this out to see if anything works on this website.  I am unsure how to move past this issue.  I have tried many options.  I might have to implement my own word vectorizor, which I already have in another project, but I think I might have to consider some more edge cases with that function. such as what happens when there's a word in the input that is not in the vocabulary set... I think I will just have the function not do anything.